Move FsProg to Prog
