+++ lang/core/src/lib.rs
@@ -1,3 +1,2 @@
 pub mod syntax;
-pub mod syntax_var;
 pub mod traits;
