--- lang/core/src/syntax_var/mod.rs
@@ -1,2 +1,0 @@
-pub mod program;
-pub use program::FsProg;
